+++ src/main/java/com/blackberry/bdp/kaboom/Worker.java
@@ -31,16 +31,17 @@ import org.slf4j.LoggerFactory;
 import com.blackberry.bdp.kaboom.api.KaBoomTopicConfig;
 import com.blackberry.bdp.common.conversion.Converter;
 import com.blackberry.bdp.common.jmx.MetricRegistrySingleton;
+import com.blackberry.bdp.kaboom.exception.LockNotAcquiredException;
 import com.blackberry.bdp.krackle.consumer.Consumer;
 
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.Metric;
 import com.codahale.metrics.MetricFilter;
 import com.codahale.metrics.Meter;
+import java.io.IOException;
 import java.nio.charset.Charset;
 import org.apache.curator.framework.recipes.cache.NodeCache;
 import org.apache.curator.framework.recipes.cache.NodeCacheListener;
-import org.apache.curator.framework.recipes.locks.InterProcessMutex;
 import org.apache.zookeeper.data.Stat;
 
 public final class Worker extends AsyncAssignee implements Runnable {
@@ -86,7 +87,7 @@ public final class Worker extends AsyncAssignee implements Runnable {
 	private KaBoomTopicConfig topicConfig;
 	private WorkerShift previousShift = null;
 	private WorkerShift currentShift;
-	private InterProcessMutex lock;
+	private NodeCache nodeCache;
 
 	static {
 		MetricRegistrySingleton.getInstance().getMetricsRegistry()
@@ -196,42 +197,6 @@ public final class Worker extends AsyncAssignee implements Runnable {
 			 });
 	}
 
-	static {
-		MetricRegistrySingleton.getInstance().getMetricsRegistry()
-			 .register("kaboom:total:avg messages written per sec", new Gauge<Long>() {
-				 @Override
-				 public Long getValue() {
-					 long sumMsgWritten = 0;
-					 int count = 0;
-					 synchronized (workersLock) {
-						 for (Worker w : workers) {
-							 count++;
-							 sumMsgWritten += w.getMsgWrittenPerSec();
-						 }
-					 }
-					 return sumMsgWritten / count;
-				 }
-
-			 });
-	}
-
-	static {
-		MetricRegistrySingleton.getInstance().getMetricsRegistry()
-			 .register("kaboom:total:total messages written per sec", new Gauge<Long>() {
-				 @Override
-				 public Long getValue() {
-					 long sumMsgWritten = 0;
-					 synchronized (workersLock) {
-						 for (Worker w : workers) {
-							 sumMsgWritten += w.getMsgWrittenPerSec();
-						 }
-					 }
-					 return sumMsgWritten;
-				 }
-
-			 });
-	}
-
 	public Worker(StartupConfig config, String topicName, int partition) throws Exception {
 		// Set up our Synchronous Worker
 		super(config.getKaBoomCurator(),
@@ -264,7 +229,7 @@ public final class Worker extends AsyncAssignee implements Runnable {
 
 		LOG.info("[{}] worker instantiated with topic configuration version {}", partitionId, topicConfig.getVersion());
 
-		NodeCache nodeCache = new NodeCache(curator, topicConfig.getZkPath());
+		nodeCache = new NodeCache(curator, topicConfig.getZkPath());
 		nodeCache.getListenable().addListener(new NodeCacheListener() {
 			@Override
 			public void nodeChanged() throws Exception {
@@ -289,8 +254,8 @@ public final class Worker extends AsyncAssignee implements Runnable {
 			}
 
 		});
+		
 		nodeCache.start();
-
 		lagGaugeName = "kaboom:partitions:" + partitionId + ":message lag";
 		lagSecGaugeName = "kaboom:partitions:" + partitionId + ":message lag sec";
 		msgWrittenGaugeName = "kaboom:partitions:" + partitionId + ":messages written per second";
@@ -366,6 +331,8 @@ public final class Worker extends AsyncAssignee implements Runnable {
 	public void run() {
 		try {
 			aquireAssignment();
+			this.hdfsOutputPath.setWorker(this);
+			
 			try {
 				currentShift = new WorkerShift();
 				hostname = InetAddress.getLocalHost().getCanonicalHostName();
@@ -578,10 +545,17 @@ public final class Worker extends AsyncAssignee implements Runnable {
 				}
 			}
 			shutdown();
+		} catch (LockNotAcquiredException lnae) {
+			LOG.error("[{}] failed to aquire lock", getPartitionId());
 		} catch (Exception e) {
 			LOG.error("[{}] An exception occured while setting up this worker thread", getPartitionId(), e);
 		} finally {
-			LOG.info("[{}] Worker finished after having processed {} events", partitionId, messagesWritten);
+			try {
+				nodeCache.close();
+				LOG.info("[{}] closed off the node cache listener", partitionId);
+			} catch (IOException ioe) {
+				LOG.error("[{}] failed to close off the node cache listener: ", partitionId, ioe);
+			}			
 			try {
 				releaseAssignment();
 			} catch (Exception ex) {
@@ -590,6 +564,7 @@ public final class Worker extends AsyncAssignee implements Runnable {
 			synchronized (workersLock) {
 				workers.remove(this);
 			}
+			LOG.info("[{}] Worker finished after having processed {} events", partitionId, messagesWritten);
 		}
 	}
 
@@ -810,10 +785,6 @@ public final class Worker extends AsyncAssignee implements Runnable {
 		return lag_sec;
 	}
 
-	public long getMsgWrittenPerSec() {
-		return messagesWritten / ((System.currentTimeMillis() - startTime) / 1000);
-	}
-
 	public String getPartitionId() {
 		return partitionId;
 	}
@@ -834,4 +805,7 @@ public final class Worker extends AsyncAssignee implements Runnable {
 		return pong;
 	}
 
+	public void setPong(Boolean val) {
+		this.pong = val;
+	}
 }